requirements.txt
azure-core==1.8.1
azure-cosmos==4.1.0
azure-functions==1.3.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
requests==2.24.0
six==1.15.0
urllib3==1.25.10
